Check timeline semaphore submit info is properly ignored
Add a test that does not use timeline semaphores but includes a
VkTimelineSemaphoreSubmitInfo structure in the submit info chain.

As per the spec, the wait and signal semaphore count in the extension
structure does not need to match the ones in the general structure, so
the test uses a count of 1 and a single value for both cases.

In the general submit info structure, the semaphore count is much
larger, so a driver could potentially attempt to read beyond the end of
the values array if it incorrectly takes it into account.

See VUID-VkSubmitInfo-pNext-03240 and others.

New test:
dEQP-VK.synchronization.timeline_semaphore.misc.ignore_timeline_semaphore_info
dEQP-VKSC.synchronization.timeline_semaphore.misc.ignore_timeline_semaphore_info

Components: Vulkan
VK-GL-CTS issue: 5958

Change-Id: I0830b8f0589e2bf4295681fc75e72bb6fa7523ca