Fix segfault at XCloseDisplay on Linux X11
Segfault occurs at XCloseDisplay when a hook from
libXext is called after it's unloaded. The cause
is the misordered destruction of the m_display and
m_library private members of the NativeDisplay class.
This change swaps the order of declaration (and thus
implicit destruction) of those private members.

Components: OpenGL

VK-GL-CTS issue: 5973

Change-Id: I1306b278c7ee346e37b69f3290453db30690a519