Increase maintenance10 resolve thresholds for D32 and D24
In practice, we detected that the original threshold here (0.000000075f)
was not enough in some cases. We need to take into account that the
reference value is calculated using floats, which have their own
precission issues, and the spec does not specify how the implementation
calculates the average. If the implementation is storing the sample
values first as D24s and sampling them later, it's losing precission
already in that step, on top of the average and final store.

Increasing the threshold to 2 units for D24 and D32 is still reasonable
and allows some implementations to pass.

Affects:
dEQP-VK.pipeline.*.multisample.m10_resolve.*depth*

Components: Vulkan
VK-GL-CTS issue: 5737

Change-Id: I97e8b1fd4b3dccf7109bbebbe90ef7a68ad55e52