Fix several issues in tessellation/geometry passthrough tests
The main issue is an OOB read when the tessellation primitive is a
triangle. The tessellation control shader declared to be emitting 4
vertices as the output, but the input patch size was 3 vertices. As the
tessellation control shader is invoked once per output vertex, and the
shader used gl_InvocationID to read the gl_in array, the last access was
out of bounds.

The second issue is a support check being run in at iterate() time
instead of at checkSupport() time, slowing down execution for
implementations not supporting geometry/tessellation.

The third issue is result images being logged even when the test passes.

Affects:
dEQP-VK.tessellation.geometry_interaction.*

VK-GL-CTS issue: 6067
Components: Vulkan

Co-authored-by: Lorenzo Rossi <snowycoder@gmail.com>
Change-Id: I03c59b6f0fb4f1173c18ba19313563eb28ab9c52