Fill FMA test input buffers at iterate time
The standard pattern of the spirv_assembly tests is to create and fill
buffers during test construction, then copy those buffers into GPU
buffers at iterate time, when the test is being run. This is acceptable
for small tests but fails quickly when large buffers are needed.

Instead of allocating test input data during FMA test construction, have
it allocated on-demand when the tests are run, so that it can be copied
directly into the GPU buffer, without a long-lasting copy on the CPU.

Components: Vulkan
Affects: dEQP-VK.spirv_assembly.instruction.*.opfma.*

Change-Id: I8bc1a5eadca9073e6c5e13355ac24084d877cdc5