Split resource creation into buffer + resource steps
The untyped pointer tests fill a buffer with data and create a Resource
from it. Sometimes, when using the data as push constants, the buffer is
needed but the resource is not. Rearrange the code so that we do not
need to create Resources that we don't need and then query buffers back
from them.

Components: Vulkan
Affects: [No functional change]
         dEQP-VK.spirv_assembly.instruction.compute.untyped_pointers.*

Change-Id: I8aa1ad0bc6a3dae4aa87e7184d302005c9737e8f