Add an UnitializedBuffer specialization of BufferInterface
Various tests are not using the data in the output Resource but have
been allocating storage for it anyway in the test spec. These specs live
much longer than the execution of the test and the data is never looked
at, even while the test is running.

UninitializedBuffer is an implementation of BufferInterface that doesn't
need to store the data. When the data is queried the result is left
uninitialised.

Components: Vulkan
Affects: dEQP-VK.spirv_assembly.*.image_sampler.*
         dEQP-VK.spirv_assembly.*.opfma.*
         dEQP-VK.spirv_assembly.*.16bit_storage.*

Change-Id: I704162655ab965d8537795f548038a9dbc8eab9c