Fix storage class mismatch with some VK_NV_raw_access_chains tests
Some of the tests mixed PhysicalStorageBuffer and StorageBuffer storage
classes, which is a violation of the rule in OpRawAccessChainNV
that requires its Storage Class operand be the same as the
Storage Class of Base.

This CL also adds back the explicit alignment, even for types that are
considered naturally aligned.

Affects:

dEQP-VK.spirv_assembly.instruction.*.raw_access_chain.*

Components: Vulkan

VK-GL-CTS issue: 5890
VK-GL-CTS issue: 5888

Change-Id: I1edaef3eab6c1ed371448c263d0bc1bad4f4515f