Avoid host visible allocations for buffers unused on the host
It looks like someone took extra care to make the buffer used on the
host be host-cached but not coherent with invalidate-before-read
calls.

Yet the other buffer only used on the device (by the
CmdBufferFill/CmdDraw with interlock) is still allocated to be in host
visible memory. We might as well make it device local.

Component: Vulkan
VK-GL-CTS issue: 6136

Affects: dEQP-VK.fragment_shader_interlock.basic.*

Change-Id: I5f676aee1b760efd77ce18ceee52e1a7ea0b2a30