Fix validation errors in DGC tests
This commit fixes the following validation errors:

* When using render pass wrappers from the pipeline construction utils
  combined with dynamic rendering, and a color attachment with a clear
  operation, the wrapper was issuing an empty render pass first with the
  clear operation for those attachments that needed it, followed by
  starting a new render pass with a load operation. The synchronization
  validation module was detecting a read-after-write hazard because
  writes from the first clear were not synchronized with loads from the
  start of the second render pass.

To solve this issue, we need to insert a barrier between the clear
operation and the start of the second render pass. However, this code
being in the pipeline construction utils means that potentially a lot of
tests could be affected.

* The tests checking the combination of DGC with descriptor buffers and
  push descriptors were not using the API correctly when
  bufferlessPushDescriptors is true. This corner of the API is not very
  well defined, but recent changes made it illegal to try to obtain the
  set layout size for the set that uses push descriptors.

We adapted the code to avoid using, binding and setting offsets for the
buffer in that situation, and to use a placeholder size for cases where
bufferlessPushDescriptors is false, expecting the implementation to
return the actual size in the buffer memory requirements for the
"scratch" buffer.

* Several tests using DGC and shader objects were not using the same
  descriptor set layouts or push constant ranges for all shader stages,
  which is forbidden by the spec.

Tests modified to use the same information in all stages.

Affects:
dEQP-VK.dgc.ext.*
(Potentially dEQP-VK.pipeline.* and others)

VK-GL-CTS issue: 6150
Components: Vulkan

Change-Id: I96ab163a392bba7aa7024810ff8a2f6fef6b05b2