Use the max GOP frame count for generated AV1 sequences
If the input file does not exist on disk, the input file used by the AV1
tests is generated with the GOPDef::frameCount of each config. The name
of the input file is not unique beteen different GOP parameters, so
there exists a race condition when multiple test cases have input
parameters that differ only by the frame count are run at the same time.

Generate the sequences with the max GOP size used, so that they are
consistent between runs.

VK-GL-CTS issue: 6166

Components: Vulkan

Affects:
dEQP-VK.video.encode.av1.*

Change-Id: I8639918166be8e5d58a35e80f8624ac42ebd1a48