Fix lod level iteration in vktSparseResourcesImageAlignedMipSize
The test creates an image with just 1 mip level, because getMipmapCount
returns 1 because isYCbCrFormat is true for this format. The driver
returns imageMipTailFirstLod = 1. When
sparseProperties.residencyAlignedMipSize is true, the test calculates a
LOD, and checks aspectRequirements.imageMipTailFirstLod against the
calculated value.

The test doesn't take into account the number of mip levels in the
image, which causes it to fail.

Affects:
dEQP-VK.sparse_resources.aligned_mip_size.*

Components: Vulkan
VK-GL-CTS issue: 6203

Co-authored-by: Timur Kristóf <timur.kristof@gmail.com>
Change-Id: I7a758be51820e1c62a04b956e39e6d682a508fa5