Remove sparse sub-block memory offset assumption
In this test, a sparse image is created and is fully bound
to a first memory region (memory1) as one unique sparse block.
The image is then filled with a given color (color1). The
same image is then fully bound to another distinct memory location
(memory2) and filled with another color (color2).

A block (non zero offset) of the image is then bound back to
an arbitrary region of memory1 and the content of the image is
dumped. The test expects to have the image filled with
color2, except for the region bound back to memory1, that should be
filled with only color1.

According to the specification

No specific mapping of image region to memory offset is defined,
i.e. the location that each texel corresponds to in memory is
implementation-dependent.

When binding back a sparse block of the image back to memory1,
it should be in such a way the content of memory 1 is guaranteed
to be what that sparse block contained to avoid reading
uninitialized values.

In fact, when the image size is not a multiple of
`VkSparseImageFormatProperties::imageGranularity`, there are some sparse
blocks whose underlying bound memory bytes will not be fully initialized.
Since the mapping of the image texels in memory is
implementation dependant applications should not assume where
these unititialized bytes should be in memory.

Since there is no way to query at which memory offset a sub-block
from a bigger sparse block will be placed in memory, the
application should fully bind the image by using several
sparse blocks and bind them at specific offsets in memmory.

When binding back any block in memory in memory 1, the
application should bind it at its previous offset to avoid reading
unitialized data.

Instead of binding the full image as one unique sparse block,
the application does the following:

- Get the number of sparse blocks (whose size will be
  `VkSparseImageFormatProperties::imageGranularity`) necessary to bind the
   full image.
- Bind each image block at a specific memory offset in memory 1
- Fill image with color 1
- Bind each image block at a specific memory offset in memory 2
- Fill image with color 2
- Bind back one sparse block to memory 1 at its previous offset.
- Verify the image data.

Affected tests:
dEQP-VK.sparse_resources.image_rebind*

Component: Vulkan

VK-GL-CTS issue: 6245

Change-Id: Iae2628598e38ecab752fcb87fc3d35b94d038315