Fix reading case list files with no EOL at EOF
A simple case list file with a single test case name and no end-of-line
after it resulted in a "Duplicate test case" error. This happened
because operator bool on a stream does not take the eofbit into account.

The best option is using the good() method in the stream instead. See
table at the bottom here:

  https://en.cppreference.com/w/cpp/io/ios_base/iostate

Components: Framework
VK-GL-CTS issue: 6237

Change-Id: I93e0f0f7a44425a99a814d508d864f04c7323882